Implement dropping a unique key
This is a bit trickier than it first sounds, because in Postgres the
constraints get a name, and can only be dropped by that name. The name
relates to the constraint at the time it was created, so for example if
the table was renamed then the constraint is still named for the old
table. Further, nothing says Postgres won't change this naming
convention in the future!

We could in principle introduce a scheme for naming the keys in this
module, and keep track of the names they would have been given at their
introduction. However, that's a bit fragile. Instead, use the Postgres
information schema to resolve the correct name. That's hopefully a bit
less fragile!

This did, however, entail making the connection available to the code
generator, so it can look up the constraint name; that in turn means one
needs to make sure that the migrations are generated and run in
lockstep, so bringing a database up over many migrations will not run
into the problem of the constraint being missing. There's an edge case
here where somebody adds and drops the same constraint in the same
migration, which would be silly, so for now this is left to produce an
error if anybody tries.